Module imagefmt::png

Structs

ExtChunk
PNG extension chunk.

Functions

read
Reads an image and converts it to requested format.

read_chunks
Like png::read but also returns the requested extension chunks.

read_info
Returns width, height and color type of the image.

write
Writes an image and converts it to requested color type.

write_chunks
Like png::write but also writes the given extension chunks.